+++ xclim/indicators/land/_snow.py
@@ -1,20 +1,13 @@
+from xclim import indices as xci
 from xclim.core.cfchecks import check_valid
-from xclim.core.indicator import Daily
-from xclim.core.units import declare_units
-from xclim.core.utils import wrapped_partial
-from xclim.indices import (
-    continuous_snow_cover_end,
-    continuous_snow_cover_start,
-    generic,
-    snd_max_doy,
-    snow_cover_duration,
-)
+from xclim.core.indicator import Daily, Daily2D
 
 __all__ = [
     "snow_cover_duration",
     "continuous_snow_cover_start",
     "continuous_snow_cover_end",
     "snd_max_doy",
+    "snow_melt_we_max",
 ]
 
 
@@ -27,7 +20,36 @@ class SnowDepth(Daily):
 class SnowCover(Daily):
     @staticmethod
     def cfcheck(snc):
-        check_valid(snc, "standard_name", "surface_snow_area_fraction ")
+        check_valid(snc, "standard_name", "surface_snow_area_fraction")
+
+
+class SnowWaterEq(Daily):
+    @staticmethod
+    def cfcheck(swe):
+        check_valid(
+            swe,
+            "standard_name",
+            [
+                "liquid_water_content_of_surface_snow",
+                "liquid_water_content_of_snow_layer",
+            ],
+        )
+
+
+class SWEPr(Daily2D):
+    @staticmethod
+    def cfcheck(swe, pr):
+        check_valid(
+            swe,
+            "standard_name",
+            [
+                "liquid_water_content_of_surface_snow",
+                "liquid_water_content_of_snow_layer",
+            ],
+        )
+        check_valid(
+            pr, "standard_name", ["precipitation_flux", "lwe_precipitation_rate"]
+        )
 
 
 snow_cover_duration = SnowDepth(
@@ -35,7 +57,7 @@ snow_cover_duration = SnowDepth(
     units="days",
     long_name="Number of days with snow depth above threshold",
     description="{freq} number of days with snow depth greater or equal to {thresh}",
-    compute=snow_cover_duration,
+    compute=xci.snow_cover_duration,
 )
 
 continuous_snow_cover_start = SnowDepth(
@@ -44,7 +66,7 @@ continuous_snow_cover_start = SnowDepth(
     long_name="Start date of continuous snow cover",
     description="Day of year when snow depth is above {thresh} for {window} consecutive days.",
     units="",
-    compute=continuous_snow_cover_start,
+    compute=xci.continuous_snow_cover_start,
 )
 
 continuous_snow_cover_end = SnowDepth(
@@ -53,7 +75,7 @@ continuous_snow_cover_end = SnowDepth(
     long_name="Start date of continuous snow cover",
     description="Day of year when snow depth is above {thresh} for {window} consecutive days.",
     units="",
-    compute=continuous_snow_cover_end,
+    compute=xci.continuous_snow_cover_end,
 )
 
 snd_max_doy = SnowDepth(
@@ -63,5 +85,23 @@ snd_max_doy = SnowDepth(
     description="{freq} day of year when snow depth reaches its maximum value.",
     units="",
     _partial=True,
-    compute=snd_max_doy,
+    compute=xci.snd_max_doy,
+)
+
+snow_melt_we_max = SnowWaterEq(
+    identifier="snow_melt_we_max",
+    standard_name="change_over_time_in_surface_snow_amount",
+    var_name="{freq}_snow_melt_we_max",
+    description="{freq} maximum negative change in melt water equivalent over {window} days.",
+    units="kg m-2",
+    compute=xci.snow_melt_we_max,
+)
+
+
+melt_and_precip_max = SWEPr(
+    identifier="melt_and_precip_max",
+    var_name="{freq}_melt_and_precip_max",
+    description="{freq} maximum precipitation flux and negative change in snow water equivalent over {window} days.",
+    units="kg m-2",
+    compute=xci.melt_and_precip_max,
 )